+++ index.ts
@@ -1,557 +1,25 @@
 import express, { Express, Request, Response } from 'express';
+import index from './api/index.routes';
+import web3 from './api/web3.routes';
 import dotenv from 'dotenv';
 import path from 'path';
-import { ethers } from 'ethers';
 
 dotenv.config();
 
 const app: Express = express();
 const port = process.env.PORT;
-app.use(express.urlencoded({ extended: true }))
 
-// Load the contract ABI and address
-const contractAddress = '0x5fbdb2315678afecb367f032d93f642f64180aa3';
-const contractABI = [
-    {
-      "inputs": [],
-      "stateMutability": "nonpayable",
-      "type": "constructor"
-    },
-    {
-      "anonymous": false,
-      "inputs": [
-        {
-          "indexed": true,
-          "internalType": "address",
-          "name": "owner",
-          "type": "address"
-        },
-        {
-          "indexed": true,
-          "internalType": "address",
-          "name": "approved",
-          "type": "address"
-        },
-        {
-          "indexed": true,
-          "internalType": "uint256",
-          "name": "tokenId",
-          "type": "uint256"
-        }
-      ],
-      "name": "Approval",
-      "type": "event"
-    },
-    {
-      "anonymous": false,
-      "inputs": [
-        {
-          "indexed": true,
-          "internalType": "address",
-          "name": "owner",
-          "type": "address"
-        },
-        {
-          "indexed": true,
-          "internalType": "address",
-          "name": "operator",
-          "type": "address"
-        },
-        {
-          "indexed": false,
-          "internalType": "bool",
-          "name": "approved",
-          "type": "bool"
-        }
-      ],
-      "name": "ApprovalForAll",
-      "type": "event"
-    },
-    {
-      "anonymous": false,
-      "inputs": [
-        {
-          "indexed": true,
-          "internalType": "address",
-          "name": "previousOwner",
-          "type": "address"
-        },
-        {
-          "indexed": true,
-          "internalType": "address",
-          "name": "newOwner",
-          "type": "address"
-        }
-      ],
-      "name": "OwnershipTransferred",
-      "type": "event"
-    },
-    {
-      "anonymous": false,
-      "inputs": [
-        {
-          "indexed": true,
-          "internalType": "address",
-          "name": "from",
-          "type": "address"
-        },
-        {
-          "indexed": true,
-          "internalType": "address",
-          "name": "to",
-          "type": "address"
-        },
-        {
-          "indexed": true,
-          "internalType": "uint256",
-          "name": "tokenId",
-          "type": "uint256"
-        }
-      ],
-      "name": "Transfer",
-      "type": "event"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "address",
-          "name": "to",
-          "type": "address"
-        },
-        {
-          "internalType": "uint256",
-          "name": "tokenId",
-          "type": "uint256"
-        }
-      ],
-      "name": "approve",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "address",
-          "name": "owner",
-          "type": "address"
-        }
-      ],
-      "name": "balanceOf",
-      "outputs": [
-        {
-          "internalType": "uint256",
-          "name": "",
-          "type": "uint256"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "uint256",
-          "name": "tokenId",
-          "type": "uint256"
-        }
-      ],
-      "name": "getApproved",
-      "outputs": [
-        {
-          "internalType": "address",
-          "name": "",
-          "type": "address"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "uint256",
-          "name": "tokenId",
-          "type": "uint256"
-        }
-      ],
-      "name": "getCertificateData",
-      "outputs": [
-        {
-          "internalType": "string",
-          "name": "studentName",
-          "type": "string"
-        },
-        {
-          "internalType": "string",
-          "name": "certificateName",
-          "type": "string"
-        },
-        {
-          "internalType": "string",
-          "name": "message",
-          "type": "string"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "address",
-          "name": "owner",
-          "type": "address"
-        },
-        {
-          "internalType": "address",
-          "name": "operator",
-          "type": "address"
-        }
-      ],
-      "name": "isApprovedForAll",
-      "outputs": [
-        {
-          "internalType": "bool",
-          "name": "",
-          "type": "bool"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    },
-    {
-      "inputs": [],
-      "name": "name",
-      "outputs": [
-        {
-          "internalType": "string",
-          "name": "",
-          "type": "string"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    },
-    {
-      "inputs": [],
-      "name": "owner",
-      "outputs": [
-        {
-          "internalType": "address",
-          "name": "",
-          "type": "address"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "uint256",
-          "name": "tokenId",
-          "type": "uint256"
-        }
-      ],
-      "name": "ownerOf",
-      "outputs": [
-        {
-          "internalType": "address",
-          "name": "",
-          "type": "address"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    },
-    {
-      "inputs": [],
-      "name": "renounceOwnership",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "address",
-          "name": "to",
-          "type": "address"
-        },
-        {
-          "internalType": "string",
-          "name": "studentName",
-          "type": "string"
-        },
-        {
-          "internalType": "string",
-          "name": "certificateName",
-          "type": "string"
-        },
-        {
-          "internalType": "string",
-          "name": "message",
-          "type": "string"
-        }
-      ],
-      "name": "safeMint",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "address",
-          "name": "from",
-          "type": "address"
-        },
-        {
-          "internalType": "address",
-          "name": "to",
-          "type": "address"
-        },
-        {
-          "internalType": "uint256",
-          "name": "tokenId",
-          "type": "uint256"
-        }
-      ],
-      "name": "safeTransferFrom",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "address",
-          "name": "from",
-          "type": "address"
-        },
-        {
-          "internalType": "address",
-          "name": "to",
-          "type": "address"
-        },
-        {
-          "internalType": "uint256",
-          "name": "tokenId",
-          "type": "uint256"
-        },
-        {
-          "internalType": "bytes",
-          "name": "data",
-          "type": "bytes"
-        }
-      ],
-      "name": "safeTransferFrom",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "address",
-          "name": "operator",
-          "type": "address"
-        },
-        {
-          "internalType": "bool",
-          "name": "approved",
-          "type": "bool"
-        }
-      ],
-      "name": "setApprovalForAll",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "bytes4",
-          "name": "interfaceId",
-          "type": "bytes4"
-        }
-      ],
-      "name": "supportsInterface",
-      "outputs": [
-        {
-          "internalType": "bool",
-          "name": "",
-          "type": "bool"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    },
-    {
-      "inputs": [],
-      "name": "symbol",
-      "outputs": [
-        {
-          "internalType": "string",
-          "name": "",
-          "type": "string"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "uint256",
-          "name": "tokenId",
-          "type": "uint256"
-        }
-      ],
-      "name": "tokenURI",
-      "outputs": [
-        {
-          "internalType": "string",
-          "name": "",
-          "type": "string"
-        }
-      ],
-      "stateMutability": "view",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "address",
-          "name": "from",
-          "type": "address"
-        },
-        {
-          "internalType": "address",
-          "name": "to",
-          "type": "address"
-        },
-        {
-          "internalType": "uint256",
-          "name": "tokenId",
-          "type": "uint256"
-        }
-      ],
-      "name": "transferFrom",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-    },
-    {
-      "inputs": [
-        {
-          "internalType": "address",
-          "name": "newOwner",
-          "type": "address"
-        }
-      ],
-      "name": "transferOwnership",
-      "outputs": [],
-      "stateMutability": "nonpayable",
-      "type": "function"
-    }
-  ]
+// Read urlencoded form payloads
+app.use(express.urlencoded({ extended: true }));
 
-// Create an ethers.js provider
-const provider = new ethers.JsonRpcProvider('http://127.0.0.1:8545/');
-
-// Connect to the contract
-const contract = new ethers.Contract(contractAddress, contractABI, provider);
-
-// POST endpoint to create a certificate
-app.post('/createCertificate', async (req: Request, res: Response) => {
-  try {
-    const { studentName, certificateName, message } = req.body;
-    // Ensure you have the sender's private key (onlyOwner)
-    const senderPrivateKey = '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80'; // Replace with your private key
-
-    // Create a wallet with the sender's private key
-    const wallet = new ethers.Wallet(senderPrivateKey, provider);
-    const transaction = await contract.connect(wallet).safeMint(
-        wallet.address,
-        studentName,
-        certificateName,
-        message
-      );
-    // Estimate gas cost for the transaction
-    // const gasEstimate = await contract.estimateGas.safeMint(wallet.address, studentName, certificateName, message);
-
-    // Create a transaction
-    // const transaction = await contract.functions.safeMint(wallet.address, studentName, certificateName, message);
-    // console.log(transaction)
-    // // Wait for the transaction to be mined
-    const receipt = await transaction.wait();
-
-    // // Check if the transaction was successful
-    if (receipt.status === 1) {
-      // Certificate minted successfully
-      res.status(200).json({ message: receipt  });
-    } else {
-      // Transaction failed
-      res.status(500).json({ message: 'Failed to create certificate' });
-    }
-    res.send(studentName+ certificateName + message);
-
-  } catch (error) {
-    console.error('Error creating certificate:', error);
-    res.status(500).json({ message: 'Error creating certificate' });
-  }
-});
-
-app.get('/verifyCertificate', async (req: Request, res: Response) => {
-    try {
-      // Ensure you have the sender's private key (onlyOwner)
-      const senderPrivateKey = '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80'; // Replace with your private key
-  
-      // Create a wallet with the sender's private key
-      const wallet = new ethers.Wallet(senderPrivateKey, provider);
-      const transaction = await contract.connect(wallet).getCertificateData(2);
-     
-      const receipt = await transaction.wait();
-  
-      // // Check if the transaction was successful
-      if (receipt.status === 1) {
-        // Certificate minted successfully
-        res.status(200).json({ message: receipt });
-      } else {
-        // Transaction failed
-        res.status(500).json({ message: 'not receipt' });
-      }
-  
-    } catch (error, transaction) {
-      console.error('Error creating certificate:', transaction);
-      res.status(500).json({ message: 'Error verifiying certificate' });
-    }
-  });
-
-// app.post("/CreateCertificate",(req: Request, res: Response) => {
-
-//     const studentName: string = req.body.studentName;
-
-//     res.send(studentName)
-
-// });
-
+// Serve static content
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.get('/', (req: Request, res: Response) => {
-    res.sendFile(path.join(__dirname, '../public', 'index.html'));
-});
-
-app.get('/create', (req: Request, res: Response) => {
-    res.sendFile(path.join(__dirname, '../public', 'create.html'));
-})
-
-app.get('/verify', (req: Request, res: Response) => {
-    res.sendFile(path.join(__dirname, '../public', 'verify.html'));
-})
+// API routes
+app.use(index);
+app.use(web3);
+//app.use(auth);
 
 app.listen(port, () => {
-    console.log(`[⚡️]: Server is running at http://localhost:${port}`);
+  console.log(`[⚡️]: Server is running at http://localhost:${port}`);
 });
